+++ lib/megaphone_client/episode.rb
